--- app/src/main/java/com/cho/navi/ui/RankFragment.kt
@@ -1,28 +1,0 @@
-package com.cho.navi.ui
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import com.cho.navi.databinding.FragmentRankBinding
-
-class RankFragment : Fragment() {
-
-    private var _binding: FragmentRankBinding? = null
-    private val binding get() = _binding!!
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        _binding = FragmentRankBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
-}
