+++ app/src/main/java/com/cho/navi/ui/RecommendationFragment.kt
@@ -1,0 +1,37 @@
+package com.cho.navi.ui
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import androidx.navigation.fragment.findNavController
+import com.cho.navi.R
+import com.cho.navi.databinding.FragmentRecommendationBinding
+
+class RecommendationFragment : Fragment() {
+
+    private var _binding: FragmentRecommendationBinding? = null
+    private val binding get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        _binding = FragmentRecommendationBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        binding.ibRecommendation.setOnClickListener {
+            findNavController().navigate(R.id.action_recommendation_to_my_page)
+        }
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
