+++ app/src/main/res/drawable/ic_unselected_recommendation.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="16dp"
+    android:height="16dp"
+    android:viewportWidth="16"
+    android:viewportHeight="16">
+    <path
+        android:fillColor="#9E9E9E"
+        android:pathData="M6.095,16C5.676,16 5.318,15.839 5.019,15.516C4.721,15.193 4.571,14.805 4.571,14.352V5.968C4.571,5.748 4.613,5.538 4.696,5.339C4.778,5.14 4.889,4.965 5.029,4.814L9.162,0.364C9.352,0.172 9.578,0.055 9.838,0.014C10.099,-0.027 10.349,0.021 10.59,0.158C10.832,0.296 11.006,0.488 11.114,0.735C11.222,0.982 11.244,1.237 11.181,1.497L10.324,5.288H14.476C14.882,5.288 15.238,5.453 15.543,5.782C15.848,6.112 16,6.496 16,6.936V8.584C16,8.68 15.99,8.783 15.971,8.893C15.952,9.003 15.924,9.106 15.886,9.202L13.6,15.011C13.486,15.286 13.295,15.519 13.029,15.712C12.762,15.904 12.483,16 12.191,16H6.095ZM1.524,16C1.105,16 0.746,15.839 0.448,15.516C0.149,15.193 0,14.805 0,14.352V6.936C0,6.483 0.149,6.095 0.448,5.772C0.746,5.449 1.105,5.288 1.524,5.288C1.943,5.288 2.301,5.449 2.6,5.772C2.898,6.095 3.048,6.483 3.048,6.936V14.352C3.048,14.805 2.898,15.193 2.6,15.516C2.301,15.839 1.943,16 1.524,16Z" />
+</vector>
