+++ app/src/main/res/drawable/selector_recommendation_menu.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:drawable="@drawable/ic_selected_recommendation" android:state_selected="true" />
+    <item android:drawable="@drawable/ic_unselected_recommendation" />
+</selector>
